dashboard/dashboard-requirements.txt
streamlit==1.49.1
streamlit-autorefresh==1.0.1
plotly==6.3.0
pandas==2.3.2
google-cloud-bigquery==3.36.0
PyYAML==6.0.2
google-auth==2.40.3
db-dtypes==1.4.3

